requirements.txt
pbr
sphinxcontrib-httpdomain
jinja2
